requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.5; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
async-timeout==4.0.3; python_version >= '3.7'
asyncio==3.4.3
attrs==23.1.0; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
emoji==2.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
frozenlist==1.4.0; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
markdown-it-py==3.0.0; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
multidict==6.0.4; python_version >= '3.7'
peewee==3.16.3
pygments==2.16.1; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
rich==13.5.3; python_full_version >= '3.7.0'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==2.0.5; python_version >= '3.7'
yarl==1.9.2; python_version >= '3.7'
